결과 분석은 별도 object class로 분리
